+++ client/src/components/attendance/CohortCheckbox.jsx
